requirements/prod.txt
tox==3.22.0
flake8==3.8.4
pandas==1.2.2
numpy==1.20.1
python-decouple==3.5
python-json-logger==2.0.1